pyproject.toml
[tool.poetry]
name = "invoice extractor llm app using google gemini pro vision large image models"
version = "0.1.0"
description = ""
authors = ["Nasir Abbas <akhtarabbas.islamicreaction@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
pypdf2 = "^3.0.1"
google-generativeai = "^0.8.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
